+++ src/Commands/InstallCommand.php
@@ -32,25 +32,24 @@ class InstallCommand extends Command
         // NPM Packages...
         $this->updateNodePackages(function ($packages) {
             return [
-                '@erbelion/vite-plugin-laravel-purgecss' => '^0.3.4',
+                '@erbelion/vite-plugin-laravel-purgecss' => '^0.4.2',
                 'autoprefixer' => '^10.4.21',
                 'nk_jtb' => '^0.15.0',
-                'postcss' => '^8.5.3',
-                'sass' => '^1.86.3',
-                'vite' => '^6.2.4',
+                'postcss' => '^8.5.6',
+                'sass' => '^1.90.0',
             ] + $packages;
         });
 
         // NPM Scripts...
         $this->updateNodeScripts(function ($scripts) {
             return [
+                'log' => 'code storage/logs/laravel.log',
                 'build' => 'vite build',
-                'build1' => 'concurrently "vite build" "npm run build:css"',
                 'build:css' => 'sass resources/scss/app.scss public/css/app.css --no-source-map --style=compressed',
                 'build:staging' => 'vite build --mode=staging',
-                'housekeeping' => 'rm -rf node_modules/.vite node_modules/.vite-cache public/build',
                 'debug' => 'vite --debug',
                 'dev' => 'concurrently "vite" "npm run watch:jtb"',
+                'housekeeping' => 'rm -rf node_modules/.vite node_modules/.vite-cache public/build',
                 'nuke' => 'rm -rf node_modules vendor public/build',
                 'nuke:ps' => 'powershell -NoProfile -Command "Remove-Item -Recurse -Force node_modules, vendor, public/build"',
                 'watch:jtb' => 'sass --watch resources/scss/app.scss public/css/app.css --no-source-map --style=compressed',